+++ src/layouts/Table/components/TableBody/TableBody.styles.ts
@@ -5,7 +5,8 @@ import type { TableBodyStyleProps } from "./TableBody.types";
 export const tableBodyStyles = ({ $style }: TableBodyStyleProps): RuleSet => {
   if ($style === "form") return css``;
   return css`
-    border: 1px solid var(--primary-light);
+    outline: 1px solid var(--primary-light);
+    outline-offset: -1px;
     border-radius: var(--s-1);
     overflow: hidden;
   `;