+++ src/layouts/Table/components/TableHeader/TableHeader.types.ts
@@ -1,0 +1,23 @@
+import type { TableProps } from "../../Table.types";
+import type { FixedCssLength, StyleProps } from "@types";
+
+export type TableHeaderProps = TableProps & {
+  /**
+   * Whether the header should be stickyÅ
+   *
+   * @type {boolean}
+   * @default false
+   */
+  sticky?: boolean;
+  /**
+   * Top alignment offset, only used if `sticky` is `true`
+   *
+   * @type {FixedCssLength}
+   * @default 0
+   */
+  top?: FixedCssLength;
+};
+
+export type TableHeaderStyleProps = StyleProps<
+  Required<Pick<TableHeaderProps, "sticky" | "top">>
+>;