+++ src/layouts/Table/components/TableHeaderCell/TableHeaderCell.styles.ts
@@ -1,5 +1,7 @@
 import { css, type RuleSet } from "styled-components";
 
+import { breakpointsPx } from "@/sizes";
+
 export const tableHeaderCellStyles = (): RuleSet => css`
   text-align: left;
   padding-bottom: var(--s-2);
@@ -9,11 +11,17 @@ export const tableHeaderCellStyles = (): RuleSet => css`
     width: 100%;
     padding: var(--s-1) var(--s-2);
     min-height: var(--s-8);
-    height: max-content;
+    height: 100%;
     border: 1px solid var(--primary-medium);
     display: flex;
     flex-direction: row;
     align-items: center;
+  }
+
+  @media only screen and (max-width: ${breakpointsPx.XS}) {
+    & > div {
+      min-height: var(--s-10);
+    }
   }
 
   &:not(:first-child) > div,