+++ src/layouts/Table/components/TableRow/TableRow.types.ts
@@ -26,5 +26,7 @@ export type TableRowProps = {
 } & StyleOverrideProps;
 
 export type TableRowStyleProps = StyleProps<
-  Required<Pick<TableRowProps, "style">>
+  Required<Pick<TableRowProps, "style">> & {
+    clickable?: boolean;
+  }
 >;