+++ src/layouts/Table/components/TableRow/index.ts
@@ -1,1 +1,2 @@
 export * from "./TableRow.component";
+export type * from "./TableRow.types";